requirements.txt
pyspark==3.5.1
delta-spark==3.2.0
pyyaml
pandas
kafka-python
fastapi
uvicorn[standard]
python-dotenv
pytest
black
isort